add reset setting
